% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfp.R
\name{bfp}
\alias{bfp}
\title{Title}
\usage{
bfp(
  ssbf_obj,
  x,
  y,
  plottype,
  color_relp = TRUE,
  subset = NULL,
  aes_args = list(),
  other_args = list()
)
}
\arguments{
\item{other_args}{}
}
\value{

}
\description{
Title
}
